refactor: Update ResourceStoreEmailPasswordStore to use KeyValueStorage
