+++ config/presets/identity/email-password-interaction-policy.json
@@ -4,15 +4,15 @@
     {
       "comment": "The storage adapter that persists usernames, passwords, etc.",
       "@id": "urn:solid-server:email-password-interaction:StorageAdapter",
-      "@type": "ResourceStoreEmailPasswordStore",
-      "ResourceStoreEmailPasswordStore:_args_baseUrl": {
+      "@type": "KeyValueEmailPasswordStore",
+      "KeyValueEmailPasswordStore:_args_baseUrl": {
         "@id": "urn:solid-server:default:variable:baseUrl"
       },
-      "ResourceStoreEmailPasswordStore:_args_storagePathname": "/idp/email-password-db",
-      "ResourceStoreEmailPasswordStore:_args_saltRounds": 10,
-      "ResourceStoreEmailPasswordStore:_args_store": {
-        "@type": "PassthroughKeyValueStore",
-        "PassthroughKeyValueStore:_store": {
+      "KeyValueEmailPasswordStore:_args_storagePathname": "/idp/email-password-db",
+      "KeyValueEmailPasswordStore:_args_saltRounds": 10,
+      "KeyValueEmailPasswordStore:_args_storage": {
+        "@type": "JsonResourceStorage",
+        "JsonResourceStorage:_source": {
           "@id": "urn:solid-server:default:ResourceStore"
         }
       }
